// lib/database.types.ts
export type Database = {
  // Allows to automatically instanciate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.12 (cd3cf9e)"
  }
  public: {
    Tables: {
      channels: {
        Row: {
          created_at: string
          discord_chat_id: string
          id: string
          telegram_chat_id: string
          type: Database["public"]["Enums"]["channel_direction"]
          updated_at: string
        }
        Insert: {
          created_at?: string
          discord_chat_id?: string
          id?: string
          telegram_chat_id?: string
          type: Database["public"]["Enums"]["channel_direction"]
          updated_at?: string
        }
        Update: {
          created_at?: string
          discord_chat_id?: string
          id?: string
          telegram_chat_id?: string
          type?: Database["public"]["Enums"]["channel_direction"]
          updated_at?: string
        }
        Relationships: []
      }
      message_mappings: {
        Row: {
          discord_id: string
          telegram_id: string
        }
        Insert: {
          discord_id: string
          telegram_id: string
        }
        Update: {
          discord_id?: string
          telegram_id?: string
        }
        Relationships: []
      }
      messages: {
        Row: {
          channel_id: string
          content: string
          created_at: string
          deleted_at: string | null
          id: string
          modified_at: string
          user_id: string
        }
        Insert: {
          channel_id: string
          content: string
          created_at?: string
          deleted_at?: string | null
          id: string
          modified_at?: string
          user_id: string
        }
        Update: {
          channel_id?: string
          content?: string
          created_at?: string
          deleted_at?: string | null
          id?: string
          modified_at?: string
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "messages_channel_id_fkey"
            columns: ["channel_id"]
            isOneToOne: false
            referencedRelation: "channels"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "messages_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      users: {
        Row: {
          created_at: string
          id: string
          modified_at: string
          pfp: string | null
          platform: Database["public"]["Enums"]["platform"]
          username: string
        }
        Insert: {
          created_at?: string
          id: string
          modified_at?: string
          pfp?: string | null
          platform: Database["public"]["Enums"]["platform"]
          username: string
        }
        Update: {
          created_at?: string
          id?: string
          modified_at?: string
          pfp?: string | null
          platform?: Database["public"]["Enums"]["platform"]
          username?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      channel_direction: "two-ways" | "one-way"
      platform: "telegram" | "discord"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never